feat: Listing files in directiry, detection of Chessboard corners  and README.md file
For the purpose of file handling there is a file_handle.cpp and its corresponding header file. It will read a given directory and save a record of detected files in an instance of the class.

For camera calibration the program uses camera_calibration.cpp and its header file, which using OpenCV tool and chessboard images taken by a camera finds the chessboard corners and by using this points will find the matrices for camera calibration.